Python/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
gcloud==0.17.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
jws==0.1.3
mysql-connector-python==8.1.0
oauth2client==3.0.0
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.4.3
pyparsing==3.1.1
Pyrebase==3.0.27
python-firebase @ git+https://github.com/ozgur/python-firebase@0d79d7609844569ea1cec4ac71cb9038e834c355
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
urllib3==2.0.4
